♻️ split compact cookie codec out of store.ts
